requirements.txt
django
djangorestframework
django-cors-headers
joblib
scikit-learn
nltk
